# zip-repl/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GEF Player Statistics API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /awards:
    get:
      operationId: listAwards
      tags: [awards]
      summary: List all awards
      responses:
        "200":
          description: List of awards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Award"
    post:
      operationId: createAward
      tags: [awards]
      summary: Create award
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAwardRequest"
      responses:
        "201":
          description: Award created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Award"
  /awards/{id}:
    delete:
      operationId: deleteAward
      tags: [awards]
      summary: Delete award
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Award:
      type: object
      properties:
        id:
          type: integer
        playerId:
          type: integer
        playerName:
          type: string
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        awardedAt:
          type: string
      required:
        - id
        - playerId
        - title
        - awardedAt
    CreateAwardRequest:
      type: object
      properties:
        playerId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        awardedAt:
          type: string
      required:
        - playerId
        - title
        - awardedAt
...